extract eclp chart and hook
